Changing location of vector db code
